+++ Package.swift
@@ -8,7 +8,7 @@
 import PackageDescription
 
 let package = Package(
-    name: "R2Navigator",
+    name: "r2-navigator-swift",
     defaultLocalization: "en",
     platforms: [.iOS(.v10)],
     products: [
@@ -17,13 +17,16 @@ let package = Package(
             targets: ["R2Navigator"]),
     ],
     dependencies: [
-        .package(url: "https://github.com/scinfu/SwiftSoup.git", .exact("2.3.2")),
-        .package(name: "R2Shared", url: "https://github.com/readium/r2-shared-swift.git", .branch("develop")),
+        .package(url: "https://github.com/scinfu/SwiftSoup.git", from: "2.3.2"),
+        .package(url: "https://github.com/readium/r2-shared-swift.git", .branch("develop")),
     ],
     targets: [
         .target(
             name: "R2Navigator",
-            dependencies: ["SwiftSoup", "R2Shared"],
+            dependencies: [
+                "SwiftSoup",
+                .product(name: "R2Shared", package: "r2-shared-swift"),
+            ],
             path: "./r2-navigator-swift/",
             exclude: ["Info.plist"],
             resources: [
@@ -31,7 +34,7 @@ let package = Package(
             ]
         ),
         .testTarget(
-            name: "r2-navigator-swiftTests",
+            name: "R2NavigatorTests",
             dependencies: ["R2Navigator"],
             path: "./r2-navigator-swiftTests/",
             exclude: ["Info.plist"]