+++ r2-navigator-swift/EPUB/EPUBSpreadView.swift
@@ -306,13 +306,8 @@ class EPUBSpreadView: UIView, Loggable, PageView {
     private static let utilsScript = loadScript(named: "utils")
 
     class func loadScript(named name: String) -> String {
-        #if SWIFT_PACKAGE
-            return Bundle.module.url(forResource: "\(name)", withExtension: "js", subdirectory: "Assets/Scripts")
-                .flatMap { try? String(contentsOf: $0) }!
-        #else
-            return Bundle(for: EPUBSpreadView.self)
-                .url(forResource: "Scripts/\(name)", withExtension: "js")
-        #endif
+        return Bundle.module.url(forResource: "\(name)", withExtension: "js", subdirectory: "Assets/Scripts")
+            .flatMap { try? String(contentsOf: $0) }!
     }
     
     func loadResource(at path: String) -> String {